fix(lint): rename l -> l_bound in shields; ruff config; auto-fix + format
