--- franka.xml
@@ -1,40 +1,0 @@
-<mujoco model="franka_planar3">
-  <option timestep="0.01" gravity="0 0 0" integrator="Euler"/>
-  <size nuser_geom="1"/>
-
-  <!-- Varsayılan eklem sönümlemesi (damping) -->
-  <default>
-    <joint damping="1.0"/>
-  </default>
-
-  <visual><map znear="0.01" zfar="10"/></visual>
-
-  <worldbody>
-    <body name="base" pos="0 0 0">
-      <joint name="j1" type="hinge" axis="0 0 1" range="-3.1416 3.1416"/>
-      <geom name="link1" type="capsule" fromto="0 0 0  0.3 0 0" size="0.02" rgba="0.3 0.5 0.8 1"/>
-
-      <body name="link2" pos="0.3 0 0">
-        <joint name="j2" type="hinge" axis="0 0 1" range="-3.1416 3.1416"/>
-        <geom name="link2" type="capsule" fromto="0 0 0  0.25 0 0" size="0.018" rgba="0.3 0.6 0.6 1"/>
-
-        <body name="link3" pos="0.25 0 0">
-          <joint name="j3" type="hinge" axis="0 0 1" range="-3.1416 3.1416"/>
-          <geom name="link3" type="capsule" fromto="0 0 0  0.2 0 0" size="0.016" rgba="0.3 0.7 0.4 1"/>
-
-          <body name="ee" pos="0.2 0 0">
-            <site name="ee_site" pos="0 0 0" size="0.01" rgba="1 0 0 1"/>
-          </body>
-        </body>
-      </body>
-    </body>
-  </worldbody>
-
-  <!-- 3 eklem için hız aktüatörleri (MuJoCo 3.x: kv) -->
-  <!-- kv'yi 0.2 yaparak agresifliği azaltıyoruz -->
-  <actuator>
-    <velocity name="v1" joint="j1" kv="0.2"/>
-    <velocity name="v2" joint="j2" kv="0.2"/>
-    <velocity name="v3" joint="j3" kv="0.2"/>
-  </actuator>
-</mujoco>
